Refactor lists looping and indexing into iteration with zipping
Disclaimer: 'zip' function merges iterables into a single iterator with the shortest lenght among the iterables. Please note that this can be relevant when merging differnt lengh iterables. Then use itertools.zip_longest

The 'zip' function merges the 3 lists into a single itarator consumed by the for loop. This way there is no need to extract the current index as the three elements returned are the elements of the current iteration in their parent list. This is more efficient and idiomatic.